requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
alabaster==1.0.0
attrs==25.3.0
babel==2.17.0
blinker==1.9.0
cachelib==0.13.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
docutils==0.21.2
Flask==3.1.1
Flask-Login==0.6.3
Flask-OAuthlib==0.9.6
Flask-SQLAlchemy==3.1.1
frozenlist==1.7.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
greenlet==3.2.3
gunicorn==23.0.0
idna==3.10
imagesize==1.4.1
itsdangerous==2.2.0
Jinja2==3.1.6
libsql-client==0.3.1
MarkupSafe==3.0.2
multidict==6.4.4
oauthlib==2.1.0
packaging==25.0
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.1
python-dotenv==1.1.0
requests==2.32.4
requests-oauthlib==1.1.0
roman-numerals-py==3.1.0
rsa==4.9.1
snowballstemmer==3.0.1
Sphinx==8.2.3
sphinx-press-theme==0.8.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.41
sqlalchemy-libsql==0.1.0
typing_extensions==4.14.0
urllib3==2.4.0
Werkzeug==3.1.3
yarl==1.20.1
PyJWT==2.8.0
